Prepare for multiple levels.
Changed end game controller to end level controller.
Implemented base class for end level controller, so
that each level can implement its own end conditions.
