+++ YouVsVirus/Assets/Scripts/Components/EndLevelControllerLevel1.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Diagnostics;
+using UnityEngine;
+
+/// <summary>
+/// The end level controller for level 1.
+/// Level 1 just ends when everybody was infected, so
+/// we keep all settings from the base class.
+/// </summary>
+public class EndLevelControllerLevel1 : EndLevelControllerBase
+{
+
+}
