+++ YouVsVirus/Assets/Scripts/Components/EndLevelControllerLevel1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ae52b6270c0ad148a94f5770b491669
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
